test: stabilize session store db isolation
