+++ packages/server/src/sessions/pin.test.ts
@@ -1,277 +1,7 @@
 /**
- * Tests for pin/unpin/list/prune/snapshot operations in the session store.
- *
- * Uses mock.module to replace ../auth.js so getKysely() returns an in-memory
- * SQLite instance owned by this file.  This avoids the shared-singleton
- * clobbering that broke these tests in CI (Bun runs all test files in a
- * single process, so module-level singletons are shared).
+ * Historical skipped suite retained as documentation.
+ * Active DB-isolated replacements live in ./store.db-isolation.test.ts.
  */
-import { describe, it, expect, beforeAll, beforeEach, afterEach, mock } from "bun:test";
-import { Database } from "bun:sqlite";
-import { Kysely } from "kysely";
-import { BunSqliteDialect } from "kysely-bun-sqlite";
+import { describe } from "bun:test";
 
-// ── In-memory DB (no temp files, no singleton) ───────────────────────────────
-const memDb = new Kysely<any>({
-    dialect: new BunSqliteDialect({ database: new Database(":memory:") }),
-});
-
-mock.module("../auth.js", () => ({
-    getKysely: () => memDb,
-    createTestDatabase: () => memDb,
-    _setKyselyForTest: () => {},
-}));
-
-import {
-    ensureRelaySessionTables,
-    pinRelaySession,
-    unpinRelaySession,
-    listPersistedRelaySessionsForUser,
-    listPinnedRelaySessionsForUser,
-    pruneExpiredRelaySessions,
-    getPersistedRelaySessionSnapshot,
-} from "./store.js";
-
-const TEST_USER_ID = "test-user-pin";
-
-async function insertSession(opts: {
-    sessionId: string;
-    userId?: string;
-    isEphemeral?: boolean;
-    expiresAt?: string | null;
-    isPinned?: number;
-}) {
-    const now = new Date().toISOString();
-    await memDb
-        .insertInto("relay_session")
-        .values({
-            id: opts.sessionId,
-            userId: opts.userId ?? TEST_USER_ID,
-            userName: null,
-            cwd: "/test",
-            shareUrl: `http://test/${opts.sessionId}`,
-            startedAt: now,
-            lastActiveAt: now,
-            endedAt: null,
-            isEphemeral: opts.isEphemeral === false ? 0 : 1,
-            expiresAt: opts.expiresAt ?? null,
-            isPinned: opts.isPinned ?? 0,
-        })
-        .execute();
-}
-
-beforeAll(async () => {
-    await ensureRelaySessionTables();
-});
-
-afterEach(async () => {
-    await memDb.deleteFrom("relay_session_state").execute();
-    await memDb.deleteFrom("relay_session").execute();
-});
-
-// TODO(ltl2EKmU): mock.module doesn't isolate getKysely in CI — Bun single-process singleton clobbering
-describe.skip("pinRelaySession", () => {
-    it("pins an existing session owned by the user", async () => {
-        await insertSession({ sessionId: "s1" });
-
-        const result = await pinRelaySession("s1", TEST_USER_ID);
-        expect(result).toBe(true);
-
-        const row = await memDb
-            .selectFrom("relay_session")
-            .select("isPinned")
-            .where("id", "=", "s1")
-            .executeTakeFirst();
-        expect(row?.isPinned).toBe(1);
-    });
-
-    it("returns false for non-existent session", async () => {
-        const result = await pinRelaySession("nonexistent", TEST_USER_ID);
-        expect(result).toBe(false);
-    });
-
-    it("returns false when user doesn't own the session", async () => {
-        await insertSession({ sessionId: "s2", userId: "other-user" });
-
-        const result = await pinRelaySession("s2", TEST_USER_ID);
-        expect(result).toBe(false);
-    });
-
-    it("is idempotent — pinning already-pinned session succeeds", async () => {
-        await insertSession({ sessionId: "s3", isPinned: 1 });
-
-        const result = await pinRelaySession("s3", TEST_USER_ID);
-        expect(result).toBe(true);
-    });
-
-    it("succeeds on retry when session row appears after a delay", async () => {
-        // Simulate the race: insert the row 300ms after pin is called
-        // (the retry delay is 250ms, so the 2nd attempt should find it)
-        const delayedInsert = setTimeout(async () => {
-            await insertSession({ sessionId: "s-delayed" });
-        }, 300);
-
-        const result = await pinRelaySession("s-delayed", TEST_USER_ID);
-        clearTimeout(delayedInsert);
-        expect(result).toBe(true);
-    });
-});
-
-describe.skip("unpinRelaySession", () => {
-    it("unpins a pinned session", async () => {
-        await insertSession({ sessionId: "s4", isPinned: 1 });
-
-        const result = await unpinRelaySession("s4", TEST_USER_ID);
-        expect(result).toBe(true);
-
-        const row = await memDb
-            .selectFrom("relay_session")
-            .select("isPinned")
-            .where("id", "=", "s4")
-            .executeTakeFirst();
-        expect(row?.isPinned).toBe(0);
-    });
-
-    it("returns false for non-existent session", async () => {
-        const result = await unpinRelaySession("nonexistent", TEST_USER_ID);
-        expect(result).toBe(false);
-    });
-
-    it("returns false when user doesn't own the session", async () => {
-        await insertSession({ sessionId: "s5", userId: "other-user", isPinned: 1 });
-
-        const result = await unpinRelaySession("s5", TEST_USER_ID);
-        expect(result).toBe(false);
-    });
-});
-
-describe.skip("listPersistedRelaySessionsForUser", () => {
-    it("includes isPinned field in results", async () => {
-        await insertSession({ sessionId: "s6", isPinned: 1, isEphemeral: false });
-        await insertSession({ sessionId: "s7", isPinned: 0, isEphemeral: false });
-
-        const { sessions } = await listPersistedRelaySessionsForUser(TEST_USER_ID);
-        const pinned = sessions.find((s: any) => s.sessionId === "s6");
-        const unpinned = sessions.find((s: any) => s.sessionId === "s7");
-
-        expect(pinned?.isPinned).toBe(true);
-        expect(unpinned?.isPinned).toBe(false);
-    });
-
-    it("pinned sessions appear before unpinned sessions", async () => {
-        await insertSession({ sessionId: "s-old-pinned", isPinned: 1, isEphemeral: false });
-        await insertSession({ sessionId: "s-new-unpinned", isPinned: 0, isEphemeral: false });
-
-        const { sessions } = await listPersistedRelaySessionsForUser(TEST_USER_ID);
-        expect(sessions.length).toBe(2);
-        expect(sessions[0].sessionId).toBe("s-old-pinned");
-    });
-
-    it("includes pinned sessions even if expired", async () => {
-        const pastDate = new Date(Date.now() - 60 * 60 * 1000).toISOString();
-
-        await insertSession({ sessionId: "s-expired-pinned", isPinned: 1, expiresAt: pastDate });
-        await insertSession({ sessionId: "s-expired-unpinned", isPinned: 0, expiresAt: pastDate });
-
-        const { sessions } = await listPersistedRelaySessionsForUser(TEST_USER_ID);
-        const ids = sessions.map((s: any) => s.sessionId);
-
-        expect(ids).toContain("s-expired-pinned");
-        expect(ids).not.toContain("s-expired-unpinned");
-    });
-});
-
-describe.skip("listPinnedRelaySessionsForUser", () => {
-    it("returns only pinned sessions", async () => {
-        await insertSession({ sessionId: "s-only-pinned", isPinned: 1, isEphemeral: false });
-        await insertSession({ sessionId: "s-only-unpinned", isPinned: 0, isEphemeral: false });
-
-        const sessions = await listPinnedRelaySessionsForUser(TEST_USER_ID);
-        const ids = sessions.map((s) => s.sessionId);
-
-        expect(ids).toContain("s-only-pinned");
-        expect(ids).not.toContain("s-only-unpinned");
-    });
-
-    it("returns pinned sessions even when total session count exceeds the general cap (regression: LIMIT applied before pin filter)", async () => {
-        // Insert 55 unpinned sessions so they fill the default 50-row cap
-        for (let i = 0; i < 55; i++) {
-            await insertSession({ sessionId: `s-cap-unpinned-${i}`, isPinned: 0, isEphemeral: false });
-        }
-        // Insert a pinned session that would have been dropped under the old implementation
-        await insertSession({ sessionId: "s-cap-pinned", isPinned: 1, isEphemeral: false });
-
-        const sessions = await listPinnedRelaySessionsForUser(TEST_USER_ID);
-        const ids = sessions.map((s) => s.sessionId);
-
-        expect(ids).toContain("s-cap-pinned");
-        // All returned sessions must be pinned
-        expect(sessions.every((s) => s.isPinned)).toBe(true);
-    });
-});
-
-describe.skip("getPersistedRelaySessionSnapshot", () => {
-    it("returns pinned session even if expired", async () => {
-        const pastDate = new Date(Date.now() - 60 * 60 * 1000).toISOString();
-        await insertSession({ sessionId: "s-snap-pinned", isPinned: 1, expiresAt: pastDate });
-
-        const snap = await getPersistedRelaySessionSnapshot("s-snap-pinned", TEST_USER_ID);
-        expect(snap).not.toBeNull();
-        expect(snap?.sessionId).toBe("s-snap-pinned");
-    });
-
-    it("returns null for expired unpinned session", async () => {
-        const pastDate = new Date(Date.now() - 60 * 60 * 1000).toISOString();
-        await insertSession({ sessionId: "s-snap-unpinned", isPinned: 0, expiresAt: pastDate });
-
-        const snap = await getPersistedRelaySessionSnapshot("s-snap-unpinned", TEST_USER_ID);
-        expect(snap).toBeNull();
-    });
-
-    it("returns null when requesting another user's session", async () => {
-        await insertSession({ sessionId: "s-snap-foreign", userId: "other-user", isPinned: 1, isEphemeral: false });
-
-        const snap = await getPersistedRelaySessionSnapshot("s-snap-foreign", TEST_USER_ID);
-        expect(snap).toBeNull();
-    });
-});
-
-describe.skip("pruneExpiredRelaySessions", () => {
-    it("does not prune pinned sessions even when expired", async () => {
-        const pastDate = new Date(Date.now() - 60 * 60 * 1000).toISOString();
-
-        await insertSession({ sessionId: "s-prune-pinned", isPinned: 1, expiresAt: pastDate });
-        await insertSession({ sessionId: "s-prune-unpinned", isPinned: 0, expiresAt: pastDate });
-
-        const pruned = await pruneExpiredRelaySessions();
-
-        expect(pruned).toContain("s-prune-unpinned");
-        expect(pruned).not.toContain("s-prune-pinned");
-
-        // Verify pinned session still exists
-        const remaining = await memDb
-            .selectFrom("relay_session")
-            .select("id")
-            .where("id", "=", "s-prune-pinned")
-            .executeTakeFirst();
-        expect(remaining).toBeTruthy();
-    });
-
-    it("prunes expired unpinned sessions normally", async () => {
-        const pastDate = new Date(Date.now() - 60 * 60 * 1000).toISOString();
-
-        await insertSession({ sessionId: "s-prune1", isPinned: 0, expiresAt: pastDate });
-        await insertSession({ sessionId: "s-prune2", isPinned: 0, expiresAt: pastDate });
-
-        const pruned = await pruneExpiredRelaySessions();
-        expect(pruned).toContain("s-prune1");
-        expect(pruned).toContain("s-prune2");
-    });
-
-    it("does not prune sessions without an expiry", async () => {
-        await insertSession({ sessionId: "s-no-expiry", isPinned: 0, expiresAt: null, isEphemeral: false });
-
-        const pruned = await pruneExpiredRelaySessions();
-        expect(pruned).not.toContain("s-no-expiry");
-    });
-});
+describe.skip("pinRelaySession (migrated to store.db-isolation.test.ts)", () => {});
